SQLite does not support decimal types.

+++ backend/Core/Features/LineItem/Queries/GetMonthlyGrossRevenueListQuery.cs
@@ -5,9 +5,9 @@ namespace Core.Features.LineItem.Queries;
 
 public static class GetMonthlyGrossRevenueListQuery
 {
-    public record Request(int Year) : IRequest<List<decimal>>;
+    public record Request(int Year) : IRequest<List<double>>;
 
-    public class Handler : IRequestHandler<Request, List<decimal>>
+    public class Handler : IRequestHandler<Request, List<double>>
     {
         private readonly ILineItemRepo _lineItemRepo;
 
@@ -16,7 +16,7 @@ public static class GetMonthlyGrossRevenueListQuery
             _lineItemRepo = lineItemRepo;
         }
 
-        public async Task<List<decimal>> Handle(Request request, CancellationToken cancellationToken)
+        public async Task<List<double>> Handle(Request request, CancellationToken cancellationToken)
         {
             var grossRevenue = await _lineItemRepo.GetMonthlyGrossRevenueList(request.Year);
             return grossRevenue;
